Polish and Numpy state macros
