+++ ui/__init__.py
@@ -1,10 +1,11 @@
 ui_modules = [
     "color_def", "sv_IO_panel", "sv_examples_menu",
-    "sv_panels", "nodeview_rclick_menu", "nodeview_space_menu", "nodeview_keymaps",
+    "sv_panels", "nodeview_rclick_menu", "nodeview_space_menu",
     "monad", "sv_icons", "presets", "nodes_replacement", "sv_panel_display_nodes",
     "sv_temporal_viewers", "sv_vep_connector",
     # bgl modules
     "bgl_callback_3dview", "bgl_callback_nodeview",
     # show git info
     "development",
+    "nodeview_keymaps",
 ]
