+++ ui/nodeview_keymaps.py
@@ -78,16 +78,24 @@ def add_keymap():
 
         # Ctrl + Left Click   | Connect Temporal Viewer
         kmi = km.keymap_items.new('node.sv_temporal_viewer', 'LEFTMOUSE', 'RELEASE', ctrl=True)
-        # kmi.properties.cut_links = False
+        kmi.properties.force_stethoscope = False
+        kmi.properties.cut_links = False
         nodeview_keymaps.append((km, kmi))
 
-        # Ctrl + Right Click   | Connect Temporal Viewer
-        kmi = km.keymap_items.new('node.sv_temporal_sthetoscope', 'RIGHTMOUSE', 'RELEASE', ctrl=True)
+        # Ctrl + Shift + Left Click   | Connect Temporal Viewer with cutting links first
+        kmi = km.keymap_items.new('node.sv_temporal_viewer', 'LEFTMOUSE', 'RELEASE', ctrl=True, shift=True)
+        kmi.properties.force_stethoscope = False
+        kmi.properties.cut_links = True
+        nodeview_keymaps.append((km, kmi))
+
+        # Ctrl + Right Click   | Connect Temporal Viewer with accumulative links
+        kmi = km.keymap_items.new('node.sv_temporal_viewer', 'RIGHTMOUSE', 'RELEASE', ctrl=True)
+        kmi.properties.force_stethoscope = True
+        kmi.properties.cut_links = False
         nodeview_keymaps.append((km, kmi))
 
         # V   | Link selected nodes
         kmi = km.keymap_items.new('node.sv_node_connector', 'V', 'PRESS')
-        # kmi.properties.cut_links = True
         nodeview_keymaps.append((km, kmi))
