+++ utils/sv_node_utils.py
@@ -1,7 +1,7 @@
 # This file is part of project Sverchok. It's copyrighted by the contributors
 # recorded in the version control history of the file, available from
 # its original location https://github.com/nortikin/sverchok/commit/master
-#  
+#
 # SPDX-License-Identifier: GPL3
 # License-Filename: LICENSE
 
@@ -41,3 +41,20 @@ def scaled_dpi():
     """
     ps = bpy.context.preferences.system
     return ps.dpi * ps.pixel_size / 72
+
+def nodes_bounding_box(selected_nodes):
+    """
+    usage:
+    minx, maxx, miny, maxy = nodes_bounding_box(selected_nodes)
+    """
+    minx = +1e10
+    maxx = -1e10
+    miny = +1e10
+    maxy = -1e10
+    for node in selected_nodes:
+        minx = min(minx, node.location.x)
+        maxx = max(maxx, node.location.x + node.width)
+        miny = min(miny, node.location.y - node.height)
+        maxy = max(maxy, node.location.y)
+
+    return minx, maxx, miny, maxy
